MSF-395 - Disabling SELinux.

+++ proxy_install.sh
@@ -107,10 +107,14 @@ WantedBy=multi-user.target""" | tee "$proxy_service" >/dev/null 2>&1 && chmod 64
       selinux_config_backup="$selinux_config_path/selinux_old"
       if test -e "$selinux_config"; then
 
-          if ! mv "$selinux_config" "$selinux_config_backup"; then
+          if ! test -e "$selinux_config_backup"; then
 
-            echo "ERROR: $selinux_config could not backup"
-            exit 1
+            echo "$selinux_config: backing up"
+            if ! mv "$selinux_config" "$selinux_config_backup"; then
+
+              echo "ERROR: $selinux_config could not backup"
+              exit 1
+            fi
           fi
 
           if echo "SELINUX=disabled" > "$selinux_config" && echo "SELINUXTYPE=targeted" | tee -a "$selinux_config"; then
